Moving application code into ldm/dream/app
